Code review fixes!
