+++ Classes/ApiPollBot.js
@@ -1,8 +1,7 @@
 const fetch = require('node-fetch');
-const {triageLooksRareMessage} = require('../Utils/activityTriager');
 
 /** Abstract parent class for all API Poll Bots */
-class ApiPollBot {
+class APIPollBot {
   /**
    * Constructor
    * @param {string} apiEndpoint - Endpoint to be hitting
@@ -45,52 +44,22 @@ class ApiPollBot {
   }
 
   /**
-   * "Abstact" function each ApiBot must implement - parses endpoint response
+   * "Abstact" function each ApiBot must implement
+   * Parses endpoint response
    * @param {*} responseData - Dict parsed from API request json
    */
   handleAPIResponse(responseData) {
     console.warn('handleAPIResponse function not implemented!');
   }
-}
-
-/** API Poller for LooksRare List and Sale events */
-class LooksRareApiPollBot extends ApiPollBot {
-  /** Constructor just calls super
-   * @param {string} apiEndpoint - Endpoint to be hitting
-   * @param {number} refreshRateMs - How often to poll the endpoint (in ms)
-   * @param {*} bot - Discord bot that will be sending messages
-   */
-  constructor(apiEndpoint, refreshRateMs, bot) {
-    super(apiEndpoint, refreshRateMs, bot);
-  }
 
   /**
-   * Parses and handles LooksRare API endpoint data
-   * Only sends events that are new
-   * Response spec: https://looksrare.github.io/api-docs/#/Events/EventController.getEvents
-   * @param {*} responseData - Dict parsed from API request json
+   * "Abstact" function each ApiBot must implement
+   * Builds and sends any Discord messages
+   * @param {*} msg - Event info dict
    */
-  handleAPIResponse(responseData) {
-    let maxTime = 0;
-    for (const data of responseData.data) {
-      const eventTime = Date.parse(data.createdAt);
-
-      // Only deal with event if it is new
-      if (this.lastUpdatedTime < eventTime) {
-        triageLooksRareMessage(data, this.bot);
-      }
-
-      // Save the time of the latest event from this batch
-      if (maxTime < eventTime) {
-        maxTime = eventTime;
-      }
-    }
-
-    // Update latest time vars if batch has new latest time
-    if (maxTime > this.lastUpdatedTime) {
-      this.lastUpdatedTime = maxTime;
-    }
+  async buildDiscordMessage(msg) {
+    console.warn('buildDiscordMessage function not implemented!');
   }
 }
 
-module.exports.LooksRareApiPollBot = LooksRareApiPollBot;
+module.exports.APIPollBot = APIPollBot;
